feat: update provider for prefetch
